+++ test/pixel_prompt_test.dart
@@ -1,8 +1,0 @@
-import 'package:pixel_prompt/pixel_prompt.dart';
-import 'package:test/test.dart';
-
-void main() {
-  test('greet', () {
-    expect(greet(), "Welcome to pixel prompt!");
-  });
-}